refactor(cmd): extract agent execution and post-up config from up.go
